USD falls continues-forex weekly outlook rfxsignals
USD falls continues-forex weekly outlook rfxsignals
USD falls continues
AUDUSD follows the USD lower. New day highs.
The USD has continued to fall, and the AUSUSD is not immune. It is trading at the highs for the day and away from its 200 hour MA (green line) at 0.6900 level. The next target is up a the 0.6932-36 which is home to a number of swing levels going back to May 14th. 

USD falls continues
Looking at the changes and ranges for the major currency pairs, apart from the GBPUSD, the USD is trading near dollar low extremes (see chart below). The USD is the weakest of the major currencies today.